+++ lib/generators/rodauth/templates/app/models/account.rb
@@ -9,6 +9,7 @@ class Account < ApplicationRecord
 end
 <% else -%>
 class Account < Sequel::Model
+  include Rodauth::Rails.model
   plugin :enum
   enum :status, unverified: 1, verified: 2, closed: 3
 end