+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* Include `Rodauth::Rails.model` in generated Sequel account model as well (@janko)
+
 ## 1.4.1 (2022-05-08)
 
 * Deprecate `Rodauth::Rails::Model` constant (@janko)